pyproject.toml
[project]
name = "cgr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chemprop>=2.0.2",
    "ergochemics",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.1",
    "mlflow>=2.22.0",
    "networkx>=3.4.2",
    "numpy>=2.2.4",
    "optuna>=4.3.0",
    "optuna-integration>=4.3.0",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "rdkit>=2024.9.6",
    "scipy>=1.15.2",
    "torcheval>=0.0.7",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = []

[tool.uv.sources]
ergochemics = { git = "https://github.com/stefanpate/ergochemics.git" }
